Move keyboard shortcuts from app to library
