+++ app/src/main/kotlin/org/wordpress/aztec/demo/MainActivity.kt
@@ -27,12 +27,10 @@ import org.wordpress.android.util.AppLog
 import org.wordpress.android.util.PermissionUtils
 import org.wordpress.android.util.ToastUtils
 import org.wordpress.aztec.AztecText
-import org.wordpress.aztec.TextFormat
 import org.wordpress.aztec.picassoloader.PicassoImageLoader
 import org.wordpress.aztec.source.SourceViewEditText
 import org.wordpress.aztec.toolbar.AztecToolbar
 import org.wordpress.aztec.toolbar.AztecToolbarClickListener
-import org.wordpress.aztec.toolbar.ToolbarAction
 import org.xml.sax.Attributes
 import org.xml.sax.helpers.AttributesImpl
 import java.io.File
@@ -114,7 +112,6 @@ class MainActivity : AppCompatActivity(), OnRequestPermissionsResultCallback, Vi
     private var addPhotoMediaDialog: AlertDialog? = null
     private var addVideoMediaDialog: AlertDialog? = null
     private var mediaUploadDialog: AlertDialog? = null
-    private var dialogShortcuts: AlertDialog? = null
     private var mediaMenu: PopupMenu? = null
 
     private var mIsKeyboardOpen = false
@@ -143,150 +140,6 @@ class MainActivity : AppCompatActivity(), OnRequestPermissionsResultCallback, Vi
         }
 
         super.onActivityResult(requestCode, resultCode, data)
-    }
-
-    override fun onKeyUp(keyCode: Int, event: KeyEvent): Boolean {
-        when (keyCode) {
-            KeyEvent.KEYCODE_1 -> {
-                if (event.isAltPressed && event.isCtrlPressed) { // Heading 1 = Alt + Ctrl + 1
-                    aztec.toggleFormatting(TextFormat.FORMAT_HEADING_1)
-                    return true
-                }
-            }
-            KeyEvent.KEYCODE_2 -> {
-                if (event.isAltPressed && event.isCtrlPressed) { // Heading 2 = Alt + Ctrl + 2
-                    aztec.toggleFormatting(TextFormat.FORMAT_HEADING_2)
-                    return true
-                }
-            }
-            KeyEvent.KEYCODE_3 -> {
-                if (event.isAltPressed && event.isCtrlPressed) { // Heading 3 = Alt + Ctrl + 3
-                    aztec.toggleFormatting(TextFormat.FORMAT_HEADING_3)
-                    return true
-                }
-            }
-            KeyEvent.KEYCODE_4 -> {
-                if (event.isAltPressed && event.isCtrlPressed) { // Heading 4 = Alt + Ctrl + 4
-                    aztec.toggleFormatting(TextFormat.FORMAT_HEADING_4)
-                    return true
-                }
-            }
-            KeyEvent.KEYCODE_5 -> {
-                if (event.isAltPressed && event.isCtrlPressed) { // Heading 5 = Alt + Ctrl + 5
-                    aztec.toggleFormatting(TextFormat.FORMAT_HEADING_5)
-                    return true
-                }
-            }
-            KeyEvent.KEYCODE_6 -> {
-                if (event.isAltPressed && event.isCtrlPressed) { // Heading 6 = Alt + Ctrl + 6
-                    aztec.toggleFormatting(TextFormat.FORMAT_HEADING_6)
-                    return true
-                }
-            }
-            KeyEvent.KEYCODE_7 -> {
-                if (event.isAltPressed && event.isCtrlPressed) { // Heading 6 = Alt + Ctrl + 7
-                    aztec.toggleFormatting(TextFormat.FORMAT_PARAGRAPH)
-                    return true
-                }
-            }
-            KeyEvent.KEYCODE_8 -> {
-                if (event.isAltPressed && event.isCtrlPressed) { // Preformatted = Alt + Ctrl + 8
-//                    TODO: Add Preformatted format.
-//                    aztec.toggleFormatting(TextFormat.FORMAT_PREFORMATTED)
-                    return true
-                }
-            }
-            KeyEvent.KEYCODE_B -> {
-                if (event.isCtrlPressed) { // Bold = Ctrl + B
-                    formattingToolbar.findViewById(ToolbarAction.BOLD.buttonId).performClick()
-                    return true
-                }
-            }
-            KeyEvent.KEYCODE_D -> {
-                if (event.isCtrlPressed) { // Strikethrough = Ctrl + D
-                    formattingToolbar.findViewById(ToolbarAction.STRIKETHROUGH.buttonId).performClick()
-                    return true
-                }
-            }
-            KeyEvent.KEYCODE_H -> {
-                if (event.isAltPressed && event.isCtrlPressed) { // Shortcuts = Alt + Ctrl + H
-                    showDialogShortcuts()
-                    return true
-                }
-            }
-            KeyEvent.KEYCODE_I -> {
-                if (event.isCtrlPressed) { // Italic = Ctrl + I
-                    formattingToolbar.findViewById(ToolbarAction.ITALIC.buttonId).performClick()
-                    return true
-                }
-            }
-            KeyEvent.KEYCODE_K -> {
-                if (event.isCtrlPressed) { // Link = Ctrl + K
-                    formattingToolbar.findViewById(ToolbarAction.LINK.buttonId).performClick()
-                    return true
-                }
-            }
-            KeyEvent.KEYCODE_M -> {
-                if (event.isAltPressed && event.isCtrlPressed) { // Media = Alt + Ctrl + M
-                    formattingToolbar.findViewById(ToolbarAction.ADD_MEDIA.buttonId).performClick()
-                    return true
-                }
-            }
-            KeyEvent.KEYCODE_O -> {
-                if (event.isAltPressed && event.isCtrlPressed) { // Ordered List = Alt + Ctrl + O
-                    formattingToolbar.findViewById(ToolbarAction.ORDERED_LIST.buttonId).performClick()
-                    return true
-                }
-            }
-            KeyEvent.KEYCODE_P -> {
-                if (event.isAltPressed && event.isCtrlPressed) { // Page Break = Alt + Ctrl + P
-                    formattingToolbar.findViewById(ToolbarAction.PAGE.buttonId).performClick()
-                    return true
-                }
-            }
-            KeyEvent.KEYCODE_Q -> {
-                if (event.isAltPressed && event.isCtrlPressed) { // Quote = Alt + Ctrl + Q
-                    formattingToolbar.findViewById(ToolbarAction.QUOTE.buttonId).performClick()
-                    return true
-                }
-            }
-            KeyEvent.KEYCODE_T -> {
-                if (event.isAltPressed && event.isCtrlPressed) { // Read More = Alt + Ctrl + T
-                    formattingToolbar.findViewById(ToolbarAction.MORE.buttonId).performClick()
-                    return true
-                }
-            }
-            KeyEvent.KEYCODE_U -> {
-                if (event.isAltPressed && event.isCtrlPressed) { // Unordered List = Alt + Ctrl + U
-                    formattingToolbar.findViewById(ToolbarAction.UNORDERED_LIST.buttonId).performClick()
-                    return true
-                } else if (event.isCtrlPressed) { // Underline = Ctrl + U
-                    formattingToolbar.findViewById(ToolbarAction.UNDERLINE.buttonId).performClick()
-                    return true
-                }
-            }
-            KeyEvent.KEYCODE_X -> {
-                if (event.isAltPressed && event.isCtrlPressed) { // Code = Alt + Ctrl + X
-//                    TODO: Add Code action.
-//                    formattingToolbar.findViewById(ToolbarAction.CODE.buttonId).performClick()
-                    return true
-                }
-            }
-            KeyEvent.KEYCODE_Y -> {
-                if (event.isCtrlPressed) { // Redo  = Ctrl + Y
-                    aztec.redo()
-                    return true
-                }
-            }
-            KeyEvent.KEYCODE_Z -> {
-                if (event.isCtrlPressed) { // Undo  = Ctrl + Z
-                    aztec.undo()
-                    return true
-                }
-            }
-        }
-
-        return super.onKeyUp(keyCode, event)
     }
 
     fun insertMediaAndSimulateUpload(bitmap: Bitmap?, mediaPath: String) {
@@ -325,11 +178,11 @@ class MainActivity : AppCompatActivity(), OnRequestPermissionsResultCallback, Vi
             }
         }
 
-        Handler().post(runnable);
-        Handler().postDelayed(runnable, 2000);
-        Handler().postDelayed(runnable, 4000);
-        Handler().postDelayed(runnable, 6000);
-        Handler().postDelayed(runnable, 8000);
+        Handler().post(runnable)
+        Handler().postDelayed(runnable, 2000)
+        Handler().postDelayed(runnable, 4000)
+        Handler().postDelayed(runnable, 6000)
+        Handler().postDelayed(runnable, 8000)
 
         aztec.refreshText()
     }
@@ -775,13 +628,5 @@ class MainActivity : AppCompatActivity(), OnRequestPermissionsResultCallback, Vi
 
     override fun mediaTapped(attrs: Attributes?, naturalWidth: Int, naturalHeight: Int) {
         ToastUtils.showToast(this, "Media tapped!")
-    }
-
-    private fun showDialogShortcuts() {
-        val layout = LayoutInflater.from(this).inflate(R.layout.dialog_shortcuts, null)
-        val builder = AlertDialog.Builder(this)
-        builder.setView(layout)
-        dialogShortcuts = builder.create()
-        dialogShortcuts!!.show()
     }
 }
