+++ app/src/main/res/values/colors.xml
@@ -6,6 +6,4 @@
     <color name="blue_medium">#ff00aadc</color>
     <color name="blue_wordpress">#ff0087be</color>
 
-    <color name="background_key">#44cccccc</color>
-
 </resources>
