+++ app/src/main/res/values/dimens.xml
@@ -13,15 +13,4 @@
 
     <bool name="is_large_tablet_landscape">false</bool>
 
-    <dimen name="height_header">48dp</dimen>
-    <dimen name="height_key">32dp</dimen>
-    <dimen name="text_header">14sp</dimen>
-    <dimen name="text_item">16sp</dimen>
-    <dimen name="text_key">14sp</dimen>
-    <dimen name="margin_dialog">24dp</dimen>
-    <dimen name="margin_key">8dp</dimen>
-    <dimen name="padding_key">8dp</dimen>
-    <dimen name="radius_corner">2dp</dimen>
-    <dimen name="width_key">28dp</dimen>
-
 </resources>
