+++ aztec/src/main/res/drawable/bg_key.xml
