+++ aztec/src/main/res/layout/dialog_shortcuts.xml
