+++ aztec/src/main/res/values/styles.xml
@@ -44,4 +44,47 @@
         <item name="textColorHint">@android:color/darker_gray</item>
     </style>
 
+    <style name="Shortcut.Key.Multiple" parent="TextAppearance.AppCompat">
+        <item name="android:background">@drawable/bg_key</item>
+        <item name="android:gravity">center</item>
+        <item name="android:layout_marginLeft">@dimen/margin_key</item>
+        <item name="android:minWidth">@dimen/width_key</item>
+        <item name="android:paddingLeft">@dimen/padding_key</item>
+        <item name="android:paddingRight">@dimen/padding_key</item>
+        <item name="android:textAllCaps">true</item>
+        <item name="android:textSize">@dimen/text_key</item>
+        <item name="android:layout_centerVertical">true</item>
+    </style>
+
+    <style name="Shortcut.Key.Single" parent="TextAppearance.AppCompat">
+        <item name="android:background">@drawable/bg_key</item>
+        <item name="android:gravity">center</item>
+        <item name="android:layout_marginLeft">@dimen/margin_key</item>
+        <item name="android:minWidth">@dimen/width_key</item>
+        <item name="android:paddingLeft">@dimen/padding_key</item>
+        <item name="android:paddingRight">@dimen/padding_key</item>
+        <item name="android:textAllCaps">true</item>
+        <item name="android:textSize">@dimen/text_key</item>
+        <item name="android:layout_centerVertical">true</item>
+        <item name="android:layout_alignParentRight">true</item>
+    </style>
+
+    <style name="Shortcut.Header" parent="TextAppearance.AppCompat">
+        <item name="android:gravity">center_vertical</item>
+        <item name="android:layout_marginLeft">@dimen/margin_dialog</item>
+        <item name="android:layout_marginRight">@dimen/margin_dialog</item>
+        <item name="android:textAllCaps">false</item>
+        <item name="android:textColor">@color/blue_wordpress</item>
+        <item name="android:textSize">@dimen/text_header</item>
+    </style>
+
+    <style name="Shortcut.Item" parent="TextAppearance.AppCompat">
+        <item name="android:ellipsize">end</item>
+        <item name="android:gravity">center_vertical</item>
+        <item name="android:maxLines">1</item>
+        <item name="android:textAllCaps">false</item>
+        <item name="android:textSize">@dimen/text_item</item>
+        <item name="android:layout_alignParentLeft">true</item>
+    </style>
+
 </resources>
